requirements/tox.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --index-url=https://devpi.cmn.prd.rheinwerk.de/rheinwerk/dev/+simple --output-file=tox.txt tox.in
#
#--index-url https://devpi.cmn.prd.rheinwerk.de/rheinwerk/dev/+simple

apipkg==1.5
    # via execnet
astroid==1.6.6
    # via pylint
atomicwrites==1.3.0
    # via pytest
attrs==19.1.0
    # via pytest
colorama==0.4.4
    # via
    #   pylint
    #   pytest
coverage==4.5.1
    # via
    #   -r tox.in
    #   pytest-cov
execnet==1.6.0
    # via pytest-cache
flake8==3.5.0
    # via -r tox.in
isort==4.3.20
    # via pylint
lazy-object-proxy==1.4.1
    # via astroid
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
mock==2.0.0
    # via -r tox.in
more-itertools==5.0.0
    # via
    #   -r tox.in
    #   pytest
pbr==5.2.1
    # via mock
pep8==1.7.1
    # via pytest-pep8
pluggy==0.6.0
    # via pytest
py==1.8.0
    # via pytest
pycodestyle==2.3.1
    # via flake8
pydocstyle==2.1.1
    # via -r tox.in
pyflakes==1.6.0
    # via flake8
pylint==1.8.2
    # via -r tox.in
pytest-cache==1.0
    # via pytest-pep8
pytest-cov==2.5.1
    # via -r tox.in
pytest-django==2.9.1
    # via -r tox.in
pytest-pep8==1.0.6
    # via -r tox.in
pytest-pythonpath==0.7.2
    # via -r tox.in
pytest==3.6.2
    # via
    #   -r tox.in
    #   pytest-cache
    #   pytest-cov
    #   pytest-django
    #   pytest-pep8
    #   pytest-pythonpath
six==1.12.0
    # via
    #   astroid
    #   mock
    #   more-itertools
    #   pydocstyle
    #   pylint
    #   pytest
snowballstemmer==1.2.1
    # via pydocstyle
wrapt==1.11.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
